clarify inmanta module name vs Python package name (#17)
With modules v2 there is a difference between the inmanta module name (my_module) and the Python package name (inmanta-module-my-module). The difference exists in the inmanta-module- prefix and the underscores/dashes.

This PR updates the module template to use the inmanta module name instead of the python package name for the module directory name to minimize confusion. This way the developer will never really interact with the Python package name directly.

issue inmanta/inmanta-core#3555